fix: decomposite mainPage. Created SearchInput, Filters, Pagination and useFetchData
