+++ src/App/pages/MainPage/components/Filters/filters.module.scss
@@ -1,0 +1,3 @@
+.multiDropdown {
+  max-width: 300px;
+}
