+++ src/App/pages/MainPage/components/Pagination/pagination.module.scss
@@ -1,25 +1,23 @@
-@use "../../../styles/variables.scss" as *;
+@use "../../../../../styles/variables.scss" as *;
 
-.main {
-  margin-bottom: 100px;
+.card_list {
+  margin-top: $space_xxxl;
 
-  &__textBlock {
-    margin-top: 100px;
-    display: flex;
-    flex-direction: column;
-    gap: $space_xl;
-    align-items: center;
-
-    p {
-      max-width: 653px;
-      text-align: center;
-    }
+  &__title_text {
+    font-size: 32px;
   }
 
-  &__search_and_filter {
+  &__text_block {
     display: flex;
-    flex-direction: column;
+    align-items: center;
     gap: $space_xl;
+  }
+
+  &__grid {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+    gap: 30px;
   }
 }
 
@@ -46,8 +44,4 @@
       cursor: not-allowed;
     }
   }
-}
-
-.loader {
-  @include loader;
 }
