+++ src/App/pages/MainPage/components/SearchInput/Index.ts
@@ -1,0 +1,1 @@
+export { default } from './SearchInput';
